Moved some code around.
