+++ include/scran/dimensionality_reduction/blocking.hpp
@@ -6,6 +6,8 @@
 
 #include <vector>
 
+#include "utils.hpp"
+
 namespace scran {
 
 namespace pca_utils {
@@ -260,6 +262,28 @@ inline void project_sparse_matrix(const SparseMatrix& emat, Eigen::MatrixXd& pcs
                 }
             }
         });
+    }
+}
+
+template<bool rows_are_dims_>
+void clean_up_projected(Eigen::MatrixXd& proj, Eigen::VectorXd& D) {
+    // Empirically centering to give nice centered PCs, because we can't
+    // guarantee that the projection is centered in this manner.
+    if constexpr(rows_are_dims_) {
+        for (size_t i = 0, iend = proj.rows(); i < iend; ++i) {
+            proj.row(i).array() -= proj.row(i).sum() / proj.cols();
+        }
+    } else {
+        for (size_t i = 0, iend = proj.cols(); i < iend; ++i) {
+            proj.col(i).array() -= proj.col(i).sum() / proj.rows();
+        }
+    }
+
+    // Variance is a somewhat murky concept when projecting a dataset onto
+    // the PC space defined by a modified version of that dataset, so we just
+    // square it and assume that only the relative value matters.
+    for (auto& d : D) {
+        d = d * d;
     }
 }
