+++ include/scran/dimensionality_reduction/utils.hpp
@@ -30,28 +30,6 @@ inline void clean_up(size_t NC, Eigen::MatrixXd& U, Eigen::VectorXd& D) {
     return;
 }
 
-template<bool rows_are_dims_>
-void clean_up_projected(Eigen::MatrixXd& proj, Eigen::VectorXd& D) {
-    // Empirically centering to give nice centered PCs, because we can't
-    // guarantee that the projection is centered in this manner.
-    if constexpr(rows_are_dims_) {
-        for (size_t i = 0, iend = proj.rows(); i < iend; ++i) {
-            proj.row(i).array() -= proj.row(i).sum() / proj.cols();
-        }
-    } else {
-        for (size_t i = 0, iend = proj.cols(); i < iend; ++i) {
-            proj.col(i).array() -= proj.col(i).sum() / proj.rows();
-        }
-    }
-
-    // Variance is a somewhat murky concept when projecting a dataset onto
-    // the PC space defined by a modified version of that dataset, so we just
-    // square it and assume that only the relative value matters.
-    for (auto& d : D) {
-        d = d * d;
-    }
-}
-
 inline double process_scale_vector(bool scale, Eigen::VectorXd& scale_v) {
     if (scale) {
         double total_var = 0;
